Move _guess_intype to utils
