+++ zyte_api/utils.py
@@ -1,5 +1,21 @@
-# -*- coding: utf-8 -*-
+import re
+from os.path import splitext
+
 from .__version__ import __version__
+
+
+def _guess_intype(file_name, lines):
+    _, dot_extension = splitext(file_name)
+    extension = dot_extension[1:]
+    if extension in {"jl", "jsonl"}:
+        return "jl"
+    if extension == "txt":
+        return "txt"
+
+    if re.search(r'^\s*\{', lines[0]):
+        return "jl"
+
+    return "txt"
 
 
 def user_agent(library):
